Reorganize files in user dir
