+++ core/src/main/kotlin/com/arcanium/data/user/UserRepository.kt
@@ -1,6 +1,7 @@
-package com.arcanium.data
+package com.arcanium.data.user
 
-import com.arcanium.domain.model.User
+import com.arcanium.domain.Resource
+import com.arcanium.domain.user.model.User
 
 interface UserRepository {
     suspend fun loginWithEmailAndPassword(email: String, password: String): Resource<User, Exception>
