+++ core/src/main/kotlin/com/arcanium/domain/Resource.kt
@@ -1,4 +1,4 @@
-package com.arcanium.data
+package com.arcanium.domain
 
 sealed class Resource<T, M>(
     open val data: T?,
