+++ core/src/main/kotlin/com/arcanium/domain/user/model/User.kt
@@ -1,4 +1,4 @@
-package com.arcanium.domain.model
+package com.arcanium.domain.user.model
 
 data class User(
     val username: String
